Move errors.py outside of the models dir
