+++ cli/src/pcluster/api/errors.py
